requirements.txt
requests
pandas 
rich
pyfiglet
numpy
gunicorn
pickle
streamlit 
pillow
colorama
wordcloud
matplotlib
streamlit_lottie